total

+++ be/Domain/Models/Lecture.cs
@@ -2,11 +2,11 @@ using App.Data;
 
 namespace App.Domain.Models
 {
-     public class Lecture
+    public class Lecture
     {
         public int Id { get; set; }
         public string LectureTitle { get; set; } = null!;
-        public double LectureDuration { get; set; }
+        public int LectureDuration { get; set; }
         public string? LectureUrl { get; set; }
         public bool IsPreviewFree { get; set; }
         public int LectureOrder { get; set; }
